[1LP][RFR] Implement ComputeRateCollection & StorageRateCollection

+++ cfme/tests/intelligence/reports/test_validate_chargeback_report.py
@@ -29,7 +29,6 @@ import pytest
 from wrapanapi import VmState
 
 import cfme.intelligence.chargeback.assignments as cb
-import cfme.intelligence.chargeback.rates as rates
 from cfme import test_requirements
 from cfme.base.credential import Credential
 from cfme.cloud.provider import CloudProvider
@@ -49,9 +48,6 @@ pf2 = ProviderFilter(classes=[SCVMMProvider], inverted=True)  # SCVMM doesn't su
 
 pytestmark = [
     pytest.mark.tier(2),
-    pytest.mark.meta(blockers=[BZ(1511099, forced_streams=['5.9', '5.8'],
-                                  unblock=lambda provider: not provider.one_of(GCEProvider)),
-                               ]),
     pytest.mark.provider(gen_func=providers, filters=[pf1, pf2], scope='module'),
     pytest.mark.usefixtures('has_no_providers_modscope', 'setup_provider_modscope'),
     test_requirements.chargeback,
@@ -438,7 +434,9 @@ def chargeback_report_default(appliance, vm_ownership, assign_default_rate, prov
     report.queue(wait_for_finish=True)
 
     yield list(report.saved_reports.all()[0].data.rows)
-    report.delete()
+
+    if report.exists:
+        report.delete()
 
 
 @pytest.fixture(scope="module")
@@ -466,32 +464,46 @@ def chargeback_report_custom(appliance, vm_ownership, assign_custom_rate, provid
     report.queue(wait_for_finish=True)
 
     yield list(report.saved_reports.all()[0].data.rows)
-    report.delete()
+
+    if report.exists:
+        report.delete()
 
 
 @pytest.fixture(scope="module")
-def new_compute_rate():
+def new_compute_rate(appliance):
     # Create a new Compute Chargeback rate
     try:
-        desc = 'custom_' + fauxfactory.gen_alphanumeric()
-        compute = rates.ComputeRate(description=desc,
-                    fields={'Used CPU':
-                            {'per_time': 'Hourly', 'variable_rate': '3'},
-                            'Used Disk I/O':
-                            {'per_time': 'Hourly', 'variable_rate': '2'},
-                            'Used Memory':
-                            {'per_time': 'Hourly', 'variable_rate': '2'}})
-        compute.create()
-        if not BZ(1532368, forced_streams=['5.9']).blocks:
-            storage = rates.StorageRate(description=desc,
-                    fields={'Used Disk Storage':
-                            {'per_time': 'Hourly', 'variable_rate': '3'}})
-            storage.create()
-        yield desc
-    finally:
-        compute.delete()
-        if not BZ(1532368, forced_streams=['5.9']).blocks:
-            storage.delete()
+        desc = 'cstm_' + fauxfactory.gen_alphanumeric()
+        compute = appliance.collections.compute_rates.create(
+            description=desc,
+            fields={
+                'Used CPU': {'per_time': 'Hourly', 'variable_rate': '3'},
+                'Used Disk I/O': {'per_time': 'Hourly', 'variable_rate': '2'},
+                'Used Memory': {'per_time': 'Hourly', 'variable_rate': '2'}
+            }
+        )
+        storage = appliance.collections.storage_rates.create(
+            description=desc,
+            fields={
+                'Used Disk Storage': {'per_time': 'Hourly', 'variable_rate': '3'}
+            }
+        )
+    except Exception as ex:
+        pytest.fail(
+            'Exception while creating compute/storage rates for chargeback report tests. {}'
+            .format(ex)
+        )
+
+    yield desc
+
+    for entity in [compute, storage]:
+        try:
+            entity.delete_if_exists()
+        except Exception as ex:
+            pytest.fail(
+                'Exception while cleaning up compute/storage rates for chargeback report tests. {}'
+                .format(ex)
+            )
 
 
 @pytest.mark.rhv3
